fix: deleting the shooting.rs file containing the non necessary point counting and fixing some documentation error in shooting tools
